Re-implement ModifyReadOnlyEntityException
