+++ src/Controller/GitSourceController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Controller;
 
 use App\Entity\GitSource;
+use App\ErrorResponse\ModifyReadOnlyEntityError;
 use App\Exception\DuplicateObjectException;
 use App\Exception\EmptyEntityIdException;
 use App\Exception\ModifyReadOnlyEntityException;
@@ -44,8 +45,10 @@ readonly class GitSourceController
     {
         if (null !== $source->getDeletedAt()) {
             throw new ModifyReadOnlyEntityException(
-                $source->getIdentifier()->getId(),
-                $source->getIdentifier()->getType(),
+                new ModifyReadOnlyEntityError(
+                    $source->getIdentifier()->getId(),
+                    $source->getIdentifier()->getType(),
+                )
             );
         }
